+++ tests/app/forms/handlers/test_field_handler.py
@@ -1,14 +1,14 @@
 from wtforms import validators
 
 from app.data_model.answer_store import AnswerStore
-from app.forms.handlers.field_handler import FieldHandler
+from app.forms.handlers.string_handler import StringHandler
 from app.validation.validators import ResponseRequired
 
 
 def test_get_mandatory_validator_optional():
     answer = {'mandatory': False}
 
-    text_area_handler = FieldHandler(answer, None, AnswerStore(), {})
+    text_area_handler = StringHandler(answer, None, AnswerStore(), {})
     validate_with = text_area_handler.get_mandatory_validator('MANDATORY_TEXTFIELD')
 
     assert isinstance(validate_with[0], validators.Optional)
@@ -17,7 +17,7 @@ def test_get_mandatory_validator_optional():
 def test_get_mandatory_validator_mandatory():
     answer = {'mandatory': True}
 
-    text_area_handler = FieldHandler(
+    text_area_handler = StringHandler(
         answer,
         {'MANDATORY_TEXTFIELD': 'This is the default mandatory message'},
         AnswerStore(),
@@ -40,7 +40,7 @@ def test_get_mandatory_validator_mandatory_with_error():
         },
     }
 
-    text_area_handler = FieldHandler(
+    text_area_handler = StringHandler(
         answer,
         {'MANDATORY_TEXTFIELD': 'This is the default mandatory message'},
         AnswerStore(),
